project/requirements.txt
pyttsx3==2.90
openai==0.27.0
secure-smtplib==0.1.1
pillow==10.0.0
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.4
numpy==1.24.3
requests==2.31.0
python-dateutil==2.8.2
markdown==3.4.3
google-calendar-api==0.2.0
google-mail-api==0.2.0
python-docx==0.8.11
python-pptx==0.6.21
amadeus==8.1.0
googlemaps==4.10.0